Tidied up to suppress Eclipse nagging
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/FOP_0-20-0_Alt-Design@197445 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/truetype/TTFFile.java
@@ -380,7 +380,7 @@ public class TTFFile extends AbstractLogEnabled {
         ansiIndex = new java.util.HashMap();
         for (int i = 32; i < Glyphs.WINANSI_ENCODING.length; i++) {
             Integer ansi = new Integer(i);
-            Integer uni = new Integer((int)Glyphs.WINANSI_ENCODING[i]);
+            Integer uni = new Integer(Glyphs.WINANSI_ENCODING[i]);
 
             List v = (List)ansiIndex.get(uni);
             if (v == null) {
@@ -515,7 +515,7 @@ public class TTFFile extends AbstractLogEnabled {
      * @return int The CapHeight
      */
     public int getCapHeight() {
-        return (int)convertTTFUnit2PDFUnit(capHeight);
+        return convertTTFUnit2PDFUnit(capHeight);
     }
 
     /**
@@ -523,7 +523,7 @@ public class TTFFile extends AbstractLogEnabled {
      * @return int The XHeight
      */
     public int getXHeight() {
-        return (int)convertTTFUnit2PDFUnit(xHeight);
+        return convertTTFUnit2PDFUnit(xHeight);
     }
 
     /**
@@ -575,10 +575,10 @@ public class TTFFile extends AbstractLogEnabled {
      */
     public int[] getFontBBox() {
         final int[] fbb = new int[4];
-        fbb[0] = (int)convertTTFUnit2PDFUnit(fontBBox1);
-        fbb[1] = (int)convertTTFUnit2PDFUnit(fontBBox2);
-        fbb[2] = (int)convertTTFUnit2PDFUnit(fontBBox3);
-        fbb[3] = (int)convertTTFUnit2PDFUnit(fontBBox4);
+        fbb[0] = convertTTFUnit2PDFUnit(fontBBox1);
+        fbb[1] = convertTTFUnit2PDFUnit(fontBBox2);
+        fbb[2] = convertTTFUnit2PDFUnit(fontBBox3);
+        fbb[3] = convertTTFUnit2PDFUnit(fontBBox4);
 
         return fbb;
     }
@@ -588,7 +588,7 @@ public class TTFFile extends AbstractLogEnabled {
      * @return int The LowerCaseAscent
      */
     public int getLowerCaseAscent() {
-        return (int)convertTTFUnit2PDFUnit(ascender);
+        return convertTTFUnit2PDFUnit(ascender);
     }
 
     /**
@@ -596,7 +596,7 @@ public class TTFFile extends AbstractLogEnabled {
      * @return int The LowerCaseDescent
      */
     public int getLowerCaseDescent() {
-        return (int)convertTTFUnit2PDFUnit(descender);
+        return convertTTFUnit2PDFUnit(descender);
     }
 
     /**
@@ -623,7 +623,7 @@ public class TTFFile extends AbstractLogEnabled {
     public int[] getWidths() {
         int[] wx = new int[mtxTab.length];
         for (int i = 0; i < wx.length; i++) {
-            wx[i] = (int)convertTTFUnit2PDFUnit(mtxTab[i].getWx());
+            wx[i] = convertTTFUnit2PDFUnit(mtxTab[i].getWx());
         }
 
         return wx;
@@ -635,7 +635,7 @@ public class TTFFile extends AbstractLogEnabled {
      * @return int Standard width
      */
     public int getCharWidth(int idx) {
-        return (int)convertTTFUnit2PDFUnit(ansiWidth[idx]);
+        return convertTTFUnit2PDFUnit(ansiWidth[idx]);
     }
 
     /**
@@ -1092,7 +1092,7 @@ public class TTFFile extends AbstractLogEnabled {
                             adjTab = new java.util.HashMap();
                         }
                         adjTab.put(new Integer(j),
-                                   new Integer((int)convertTTFUnit2PDFUnit(kpx)));
+                                   new Integer(convertTTFUnit2PDFUnit(kpx)));
                         kerningTab.put(iObj, adjTab);
                     }
                 }
@@ -1231,8 +1231,8 @@ public class TTFFile extends AbstractLogEnabled {
         System.out.println("Family name: " + familyName);
         System.out.println("Subfamily name: " + subFamilyName);
         System.out.println("Notice:      " + notice);
-        System.out.println("xHeight:     " + (int)convertTTFUnit2PDFUnit(xHeight));
-        System.out.println("capheight:   " + (int)convertTTFUnit2PDFUnit(capHeight));
+        System.out.println("xHeight:     " + convertTTFUnit2PDFUnit(xHeight));
+        System.out.println("capheight:   " + convertTTFUnit2PDFUnit(capHeight));
 
         int italic = (int)(italicAngle >> 16);
         System.out.println("Italic:      " + italic);
@@ -1245,10 +1245,10 @@ public class TTFFile extends AbstractLogEnabled {
         System.out.println();
         System.out.println("Ascender:    " + convertTTFUnit2PDFUnit(ascender));
         System.out.println("Descender:   " + convertTTFUnit2PDFUnit(descender));
-        System.out.println("FontBBox:    [" + (int)convertTTFUnit2PDFUnit(fontBBox1)
-                           + " " + (int)convertTTFUnit2PDFUnit(fontBBox2) + " "
-                           + (int)convertTTFUnit2PDFUnit(fontBBox3) + " "
-                           + (int)convertTTFUnit2PDFUnit(fontBBox4) + "]");
+        System.out.println("FontBBox:    [" + convertTTFUnit2PDFUnit(fontBBox1)
+                           + " " + convertTTFUnit2PDFUnit(fontBBox2) + " "
+                           + convertTTFUnit2PDFUnit(fontBBox3) + " "
+                           + convertTTFUnit2PDFUnit(fontBBox4) + "]");
     }
 
     /**
